Install playwright to tabs
