+++ packages/boilerplate/example/src/index.html
@@ -2,6 +2,7 @@
 <html lang="en">
 	<head>
     <title>StormID</title>
+	<meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <style>
         * {
             margin: 0;
