+++ packages/outliner/example/src/index.html
@@ -1,38 +1,39 @@
 <!doctype html>
 <html lang="en">
 	<head>
-    <title>StormID</title>
-    <style>
-        * {
-            margin: 0;
-            padding: 0;
-            box-sizing: border-box;
-            font-family: sans-serif;
-        }
-		body {
-        	font: 100%/1.4 "AvenirNextLTW01-Regular", Helvetica, "Lucida Grande", sans-serif;
-		}
-        html,
-        body {
-            height: 100%;
-            background: #dddddd;
-        }
-		.container {
-			margin: 24px auto;
-			max-width: 960px;
-			padding: 0 24px;
-		}
-        button {
-            padding: 10px 5px;
-        }
-        button:focus {
-            outline: 3px solid blue;
-        }
-        .no-outline * {
-            outline: 0 none !important;
-        }
-    </style>
-</head>
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <title>StormID</title>
+        <style>
+            * {
+                margin: 0;
+                padding: 0;
+                box-sizing: border-box;
+                font-family: sans-serif;
+            }
+            body {
+                font: 100%/1.4 "AvenirNextLTW01-Regular", Helvetica, "Lucida Grande", sans-serif;
+            }
+            html,
+            body {
+                height: 100%;
+                background: #dddddd;
+            }
+            .container {
+                margin: 24px auto;
+                max-width: 960px;
+                padding: 0 24px;
+            }
+            button {
+                padding: 10px 5px;
+            }
+            button:focus {
+                outline: 3px solid blue;
+            }
+            .no-outline * {
+                outline: 0 none !important;
+            }
+        </style>
+    </head>
 <body>
     <main class="container">
         <h1>Outliner example</h1>
