+++ packages/tabs/__tests__/jest/integrations.js
@@ -1,4 +1,4 @@
-import tabs from '../src';
+import tabs from '../../src';
 
 let TabSet;
