+++ packages/tabs/example/src/index.html
@@ -1,5 +1,5 @@
 <!doctype html>
-<html>
+<html lang="en">
   <head>
     <title>StormID</title>
     <style>
@@ -55,7 +55,8 @@
     </style>
   </head>
   <body>
-    <div class="container">
+    <main class="container">
+        <h1>Tabs example</h1>
         <div class="tabs js-tabs" >
             <div class="tabs__tabslist" role="tablist" data-active-index="2">
                 <a id="tab-1" class="tabs__tab js-tabs__link" href="#panel-1" role="tab">Tab 1</a>
