+++ packages/tabs/tools/webpack.config.js
